[141] Small fix to update system, pushed to v1.1

+++ Units/U_Parser.pas
@@ -97,7 +97,7 @@ implementation
 
     function updateParser.getDirectDownloadLink(swLink: string): string;
     var
-        i:       byte;
+        i:       integer;
         srcTags: iHTMLElementCollection;
         srcTagE: iHTMLElement;
         srcElem: iHTMLElement2;
